Search and filters are now fully fully functional

+++ assets/js/project_page.js
@@ -39,6 +39,9 @@
                             let newBug = newBugDom(bug);
                             $(bugTable).append(newBug);
                         }
+                    },
+                    error: function(err){
+                        console.log(err.responseText);
                     }
                 });
                 
@@ -46,8 +49,104 @@
         });
     }
 
-    let filterByLabels = function(){
+    let filterByLabel = function(){
+        let projectId = $('main').attr('id').split('-')[1];
+        let filters = $('.label-filter');
+        filterElements = [];
 
+        $(filters).change(function(){
+            let self = this;
+            let filterValue = $(self).attr('id').split('-')[1];
+
+            $('.author-filter').each(function(){
+                let self = this;
+                if($(self).is(':checked')){
+                    $(self).prop('checked', false);
+                }
+            });
+
+
+            if($(self).is(':checked')){
+                filterElements.push(filterValue);
+
+            }else if(filterElements.includes(filterValue) && !$(self).is(':checked')){
+
+                filterElements.splice(filterElements.indexOf(filterValue), 1);
+            }
+
+            if(filterElements.length == 0){
+                showFullBugList(projectId);
+            }else{
+        
+                $.ajax({
+                    type: 'GET',
+                    url: `/issue/filter/label/${projectId}`,
+                    data: {
+                        labels: filterElements 
+                    },
+                    success: function(data){
+                        let bugTable = $('table');
+
+                            bugTable.html('<tr><td>Title</td><td>Description</td><td>Labels</td><td>Author</td></tr>');
+
+                            for(bug of data.data.bugs){
+                                let newBug = newBugDom(bug);
+                                $(bugTable).append(newBug);
+                            }
+                    },
+                    error: function(err){
+                        console.log(err.responseText);
+                    }
+                });
+
+            }
+        });
+    }
+
+    let filterByAuthor = function(){
+        let projectId = $('main').attr('id').split('-')[1];
+        let filters = $('.author-filter');
+
+        $(filters).change(function(){
+            let self = this;
+            let filterValue = $(self).attr('id').split('-')[1];
+
+            $('.label-filter').each(function(){
+                let self = this;
+                if($(self).is(':checked')){
+                    $(self).prop('checked', false);
+                }
+            });
+
+            $(filters).not(this).prop('checked', false);
+
+            if($(self).is(':checked')){
+                
+                $.ajax({
+                    type: 'GET',
+                    url: `/issue/filter/author/${projectId}`,
+                    data: {
+                        author: filterValue
+                    },
+                    success: function(data){
+                        let bugTable = $('table');
+
+                            bugTable.html('<tr><td>Title</td><td>Description</td><td>Labels</td><td>Author</td></tr>');
+
+                            for(bug of data.data.bugs){
+                                let newBug = newBugDom(bug);
+                                $(bugTable).append(newBug);
+                            }
+                    },
+                    error: function(err){
+                        console.log(err.responseText);
+                    }
+                });
+
+            }else{
+                showFullBugList(projectId);
+            }
+        });
     }
 
     let showFullBugList = function(projectId){
@@ -82,26 +181,50 @@
                 labels += bug.labels[i] + ", ";
             }
         }
+        let user = $('main').attr('data-user');
 
-        return $(`
+        if(user && user == bug.author._id){
+            return $(`
         <tr id="bug-${bug._id}">
         <td class="bug-title">
-            <a href="/issue/delete/${bug._id}>" class="delete-bug">
-              <i class="fa-solid fa-trash-can" ></i>
-            </a>
-          ${bug.title}
+            <div>
+                <a href="/issue/${bug._id}">${bug.title}</a>
+                <a href="/issue/delete/${bug._id}>" class="delete-bug">
+                <i class="fa-solid fa-trash-can" ></i>
+                </a>
+            </div>
         </td>
         <td>
           <p class="description"> ${bug.description}</p>
         </td>
         <td class="labels">
-            ${labels}
+            <p>${labels}</p>
         </td>
         <td>
           ${bug.author.name}
         </td>
       </tr>
         `);
+        }else{
+            return $(`
+        <tr id="bug-${bug._id}">
+        <td class="bug-title">
+          ${bug.title}
+        </td>
+        <td>
+          <p class="description"> ${bug.description}</p>
+        </td>
+        <td class="labels">
+            <p>${labels}</p>
+        </td>
+        <td>
+          ${bug.author.name}
+        </td>
+      </tr>
+        `);
+        }
+
+        
     }
 
 
@@ -109,4 +232,6 @@
 
     deleteBug();
     searchBugsBySearchTerm();
+    filterByLabel();
+    filterByAuthor();
 }
